requirements.txt
# GoPiGo3 Core
gopigo3>=1.3.2

# Web Interface
Flask>=2.3.0
Flask-SocketIO>=5.3.0

# Computer Vision (optional - uncomment if using cameras)
# opencv-python>=4.8.0
# picamera2

# Sensor Libraries
smbus2>=0.4.2
RPi.GPIO>=0.7.1

# Navigation & Math
numpy>=1.24.0

# Configuration & Data
PyYAML>=6.0

# Utilities
python-dotenv>=1.0.0

